Accept comm-count regressions as a result of IfExpr node
These tests regressed because of the following line in the
'chpl__transferArray' function in ChapelArray:

```
if (a.eltType == b.type || ...)
```

The compiler transforms the OR into an if-expr whose condition is the
left-hand side of the OR expression. In this case the compiler will
insert a call to 'chpl__convertValueToRuntimeType' for 'b.type', but
will only check the static properties of the type.

When if-exprs were implemented as functions, the call would be removed
because the function would be param-folded away. Now that if-exprs lower
to CondStmts, the call is left in the tree and the compiler doesn't know
it can be removed.

The call to 'chpl__convertValueToRuntimeType' introduces a couple of
extra GETs, leading to these regressions.