Respond to reviewer feedback part 1
Don't include the new module by default, explicitly use it when I need it.  This
meant moving it to modules/packages since we don't search modules/internal when
looking for uses to resolve.

Rename _getArray to makeArrayFromPtr and move it into the module with the new
types.

Have dsiClone return the same dist, so it is a singleton class

Update tests reflecting these changes