DataFrame ADT (#9696)
[reviewed by @benharsh]

User Story:
As a Chapel Programmer, I want to use the Dataframe ADT because it best represents the tabular nature of my data while providing powerful tools for analysis. DataFrames can be initialized with or without Series, with Series being copied into a DataFrame.

Limitations:
Currently it is not possible to emit a row from a DataFrame because a row requires a variable length, variable type ADT. See #9660 for exploration of this issue.

Notes:
Included in this PR are pandas equivalents to our tests. The only significant difference is slightly different spacing when printing a DataFrame.