In mem tracking, replace sync var with pthread mutex for concurrency.
Memory tracking has been using a sync var for concurrency control on its
data structures.  Here, replace that with a pthread mutex.  The problem
was that with a sync var, using it after the tasking layer had been shut
down (for printing the mem tracking info, for example) would cause the
qthreads tasking layer to try to create a qthread to handle the sync var
ops, but the underlying memory pools had already been shut down.  Chaos
resulted.

We considered using an atomic instead of a pthread mutex, but that would
run into the same problem as the sync var if the atomics were configured
with CHPL_ATOMICS=locks.