Prevent serialization from qthread sync vars
Previously, qthreads/tasks that were blocked on a sync var got scheduled onto
whatever shepherd happened to wake them up. The implementation then relied on
work stealing to redistribute the tasks. However, we either use schedulers that
don't support stealing (nemesis) or we run with stealing disabled (distrib w/
STEAL_RATIO=0) so this resulted in all tasks getting scheduled onto the same
shephered, which effectively serializes execution for those tasks until they
complete.

This just forces qthreads/tasks to get rescheduled on their original shepherd.
Ideally we would check if the scheduler supports stealing and if stealing is
enabled, but there's currently no reliable way to detect those things for all
schedulers, and adding them is more work than I want to take on right now.

This prevents serialization from sync vars, which allows the SPMD version of
stream to achieve full performance when using a sync barrier:

| config           | performance |
| ---------------- | ----------- |
| no barrier       | 84 GB/s     |
| atomic barrier   | 84 GB/s     |
| old sync barrier | 10 GB/s     |
| sync barrier     | 84 GB/s     |

Note that this will hurt threadring's performance. Threadring just passes a
sync token between ~500 tasks one at a time, so running the tasks serially was
actually beneficial. I expect threadring to have a 2x or so slowdown, but it's
not an important benchmark and the serializing behavior was wrong.
before.