Add error message when we can't determine an iterator's yielded type
While working on ranges of enums, I was surprised to find that the
compiler generates an unclear assertion error in the case that it
can't resolve the type yielded by an iterator.  This is a very
simple fix to at least turn it into a user-facing error pointing to
the iterator in question.