Use same interface to allocate and free
Greg pointed out that it was a bad idea to use c_malloc and chpl_mem_free to
allocate and free the memory, since it was possible that c_malloc could be
defined in the future to not call down to chpl_mem_alloc.  Instead, define a
function to create the external_array instance which will call chpl_mem_alloc
and set its free function appropriately - use this new function instead in my
module code.