Re-enable testing of partial reductions over DR
This re-enables testing of cg-sparse-partred.chpl,
following the lead of #9755.

* CSimpl.chpl, which implements partial reduction for the CS layout,
is again essentially identical to DRimpl.chpl, as it was before.

* Both impls use the utility class PartRedOp, so I moved it to
utilities.chpl.

* I removed _to_unmanaged in DRimpl.chpl. The corresponding formal
is not used. When/if we start using it, we can decide what kind
of class thing to pass to it.

TODO remains: I would like to replace the two "new unmanaged" +
"delete" pairs in each impl with just "new owned". This is what
the desired semantics is. However that change gave me some
compilation errors so I reverted it.

The TODOs regarding generate() in #9755 apply here equally.

Memory Leaks: cg-sparse-partred.chpl leaks 456 bytes.

This leak is unrelated to partial reductions. cg-sparse-partred.chpl leaks
exactly the same objects as cg-sparse.chpl. They are due solely to the line:

  const MatrixSpace: sparse subdomain(DenseSpace) dmapped(new dmap(new CS()))
                   = genAIndsSorted(elemType, n, nonzer, shift);