Don't include chpl-tasks.h in standalone mode.
In PR #9736 I added an include of chpl-tasks.h to qio.h, to make up for
the removal of a nested include of that same file in something else
qio.h included.  Unfortunately I added the new chpl-tasks.h include in a
place that shadowed the include of pthreads.h in the "standalone" side
of qio.h, which changed what pthread.h declared and thus caused failures
in standalone testing.  Here, move the include of chpl-tasks.h to the
"runtime" side of qio.h and thus remove the shadowing.