Initial attempt at disabling enum->int casts for those with no values
This is an initial attempt at disabling casts from enums to ints
for enums that have no values defined for their symbols.  To be
safe, it disables all param int->enum casts.  It may be with some
effort that these could be re-enabled for enums that do have
symbols associated with them.