Convert halt for gettimeofday to an assert in DateTime
gettimeofday() can only fail if "tv or tz pointed outside the accessible
address space" or if "Timezone (or something else) is invalid". tv/tz will
always be inside the accessible address space, and we pass in null for the
timezone, so this routine can't really fail. Just convert the halt into an
assert.