Enable providing runtime layer's hwloc topology to others.
Add chpl_topo_getHwlocTopology(), a new Chapel runtime topology layer
function which returns the topology the runtime has loaded, if any.  In
addition, where we used to only load an hwloc topology in the runtime
when locModel!=flat, we now also do so when tasks==qthreads.  This is to
prepare for the next change, which will have Qthreads use this topology.