Have Qthreads use Chapel's copy of the hwloc topology.
Add Qthreads configure and Makefile support for the new affinity shim
"hwloc_via_chapel", and default to using it.

By use of linker --wrap options and hwloc init, load, and destroy
wrapper functions I've confirmed that we do and don't load an hwloc
topology appropriately in various locModel and tasking configs.  I also
used this technique along with running the performance/elliot/no-op test
to confirm that we only load the topology once rather than twice in the
Qthreads-based configs, demonstrating that Qthreads is indeed using the
runtime's copy of the topology.  Finally, I ran stream-ep on XC-16 BW28
nodes with and without QT_INFO=1 using the nightly performance settings
to confirm that Qthreads was computing the expected number of CPUs and
getting the expected performance.