Add --break-on-id and --break-on-delete-id to chpldoc
When debugging an issue with chpldoc recently, I realized that I could debug
its operations, but couldn't use --break-on-id (a developer flag with normal
chpl compilation that I have found quite helpful when I know the id to use)

This change makes it so that chpldoc allows --break-on-id and
--break-on-delete-id to work, by defining them through a macro that is used
both in the normal driver and in the chpldoc driver.  I figured
--break-on-delete-id was also useful, but that anything that would apply after
(like --break-on-resolve-id or --break-on-codegen-id) was absolutely not going
to be necessary.

Double checked on a local --lldb run for both flags both in chpldoc and chpl
compilation.  Will run a paratest since I know there is a help output test but
I don't remember where it lives