Get CPU counts from the runtime topology interface.
Change the authoritative runtime interface for CPU counts from the
"system" layer to the topology layer.  Implement new topology layer
functions that get CPU counts, and change most of the calls to the old
system-based functions so that they go to the new topology-based ones
instead.  Rename the old system-based functions, to better indicate
where they get their information and to differentiate them from the new
functions.

The only remaining calls to the old CPU count functions are in the
implementations of the new topology layer functions.  For topo=none
those will stay as they are, but for topo=hwloc those calls are just
placeholders and the real implementation will come next.