Rewrite tests that used to use enum ranges to use them again
This rewrites a number of tests that I changed in PR #9192 as
we were deprecating enum->int coercions, to make them closer to
the way they were originally written.

In writing them, I found that bulk transfer currently assumes that the
source and target index sets are identical and that this caused
problems when an array indexed by enums was assigned an array literal
(assumed to use int indices).  Fixed these problems and filed a simple
test to lock the behavior in since I hadn't hit it previously.