Avoid a remote memory region lookup for network atomics
This passes the `remote_mr` to `do_nic_amo()` instead of having `do_nic_amo()`
perform the lookup itself. The remote address must be registered in order for
us to do a network atomic, which means we're already doing a lookup beforehand.
This just saves the result of the `mreg_for_remote_addr()` that we're doing
beforehand and passes it along to `do_nic_amo()`.

Roughly speaking this transforms:

    if (mreg_for_remote_addr(obj, loc) == NULL)
      do_fork_amo(...);
    else
      do_nic_amo(...);  // was doing mreg_for_remote_addr() itself

into:

    mem_region_t* remote_mr;
    if ((remote_mr = mreg_for_remote_addr(obj, loc)) == NULL)
      do_fork_amo(...);
    else
      do_nic_amo(..., remote_mr); // _not_ doing mreg_for_remote_addr()

In practice, I don't expect there to be a noticeable performance difference
here because `mreg_for_remote_addr()` caches the most recent lookup and we're
just optimizing out a lookup that immediately follows the first one, so the
value is going to be in the cache. Still the cache lookup isn't free and
there's no reason to do a second lookup, so just avoid it.