Update handling of type constructors for nested types with initializers
A nested type's type constructor will no longer be a method in the
initializers implementation, so we need to distinguish between type
constructors for types with the same name. For example:

```chpl
class Node { }
class List {
  var n : Node; // which _type_construct_Node?
  class Node { }
}
```

This is accomplished by setting the call's baseExpr to the appropriate
type's type constructor function.

hack_resolve_types is updated to not remove a type constructor call,
otherwise we would encounter resolution issues with generic types.