Use the right index for the cpuset of cores.
Oops!  Outsmarted myself.  The OS index for core objects isn't unique;
specifically, it can be the same for two cores on different packages.
Using it led to under-counting the cores corresponding to accessible PUs
and thus cut the reported stream-ep bandwidth on XC-16 in half.  Use the
logical index instead; that is unique.