Allow for hwloc_get_proc_cpubind() not supported on macOS.
PR #9862 broke on macOS because hwloc_get_proc_cpubind() returns "not
supported" there.  Allow for that, basically ignoring CPU binding on
that platform.  This will at least get release/examples and the smoke
testing working again.  Afterward I'll look further into whether we
need to support the notion of accessible-vs-not on macOS.