Optimize the non-fetching version of do_nic_amo
Remove several branches that are only needed for fetching atomics. This
eliminates 5 branches and the routine is now small enough to mark it as inline.
If the compiler actually inlines, it should be able to fold away a sizeof
branch as well.

I don't expect any noticeable performance improvements from this, but it does
eliminate several unnecessary branches, and I expect we may want to specialize
the non-fetching case even more in the future.