Update resolveTypeConstructor to invoke specific type constructor
Prior to this commit we were relying on resolution to find the appropriate type
constructor which could lead to issues if there were types with the same name
but in different scopes. Instead simply invoke the type constructor that we
can find from the constructor's type.

Adds a new nested type test for a top-level and nested type with the same
name. Also retires the module_class_name_clash future.