Optimize post_fma_and_wait_amo to only yield every 64 attempts
Network atomics are extremely fast, and yielding seems to be more expensive
than we think it is. We can still do an initial yield to allow for comm/compute
overlap, but if we're not oversubscribed, doing a yield every attempt slows
things down.

This has a 25% boost for RA-atomics, and takes the non-fetching many-to-one
perf test from ~210 Mops/s to ~240 Mops/s. I switched the fetching AMO case
back to use the regular post_fma_and_wait that yields every attempt. Not
yielding hurt the fetching case, so I think that needs more investigation.