Check/track array memory allocation from the comm layer.
When array memory was gotten from the comm layer, we weren't doing the
memhook calls that provide for out-of-memory reporting and memory
tracking.  Add these.  As part of this it turned out that the only
practical way to avoid doing the memhook calls twice in some cases was
to change the fallback calls used when comm-layer allocation isn't done,
so they go directly to the layer implementation interface rather than to
the public memory interface.

While here, I also removed the logic that dealt with localization when
the memory came from the comm layer, because I concluded that there was
no execution path along which the localization call actually occurred.