Make 'for' loop in dsiReallocate parallel
As noted in JIRA 189 and a comment in the code, our dsiReallocate()
code contained a serial for loop that by all rights should have been
parallel.  The last time I tried to convert it over, I had failures in
testing, but I am not seeing these failures today so am exploring
making it parallel and running testing to see what might fail.