Address several items of feedback from BHarsh
The main one here was to introduce a range method chpl_idxToInt()
that uses the range's idxType rather than requiring it to be
passed in explicitly.  Note that I had to use a different name
(I used a different number of underscores) in order to avoid the
calls in the initializers refer to the method (they're not allowed
to since the object isn't created yet).