Prevent leak when array field has no type expr, but has default value
When a field in a type has an array as its default expression, like so:

record R {
  var A = [1, 2, 3];
}

the compiler would transform the corresponding formal in the default
initializer such that the formal's type-expression would be a copy of
the default-expression, with a PRIM_TYPEOF at the end:

proc init(in A : typeof([1, 2, 3]) = [1, 2, 3]);

When the wrapper helper functions were created during resolution, the
compiler would change behavior depending on whether the type-expression
actually returned a type. In such cases, the compiler would first create
an array from the type-expression's runtime type, and then assign into
it the default value:

proc init_default_a() {
  var ret = [1,2,3]; // type expression
  ret = [1,2,3]; // default value
  return ret;
}

This is done in order to preserve the expected runtime type of the
formal. Unfortunately one of the arrays would leak.

In this case, however, it is obvious to a human that there should not be
two creations of the array "[1, 2, 3]". To fix this, we can simply copy
the default expression to the type expression without inserting a
PRIM_TYPEOF. This mimics the 'hack_resolve_types' function in
normalize.cpp and avoids creating the extra array.

A result of this change is such cases no longer leak.