Oops -- minor fix to enumRanges branch I just merged
I missed that this edit wasn't in the PR I just merged.
This reflects an improvement Ben asked me to make to
turn an unsafe cast into a safe cast.  It turns out that
our code has long relied on unsigned integer wraparound
in the unsafe cast case, so the change was not possible.
I thought I'd committed this before merging, but forgot
that I had not.