Update output for non-param for loop to reflect bool range
It's unfortunate that this causes param loops over bool ranges
to behave differently than non-param loops.  I didn't realize
that the former case would even work.  I'd be OK with this as
a stepping stone towards a better world, but may look into
how hard it would be to just support the param loop case as well
now...