Clean up halts in glob() iterators
Previously, the glob iterators used a __primitive("chpl_error", ...) if there
was an unhandled error from `chpl_clob()` (a thin wrapper over `glob()`). When
calling `glob()` with no flags like we do, the only possible error codes are
`GLOB_NOMATCH` and `GLOB_NOSPACE`. `GLOB_NOMATCH` is innocuous and will just
result in an empty iterator. `GLOB_NOSPACE` occurs if the `malloc()` inside of
glob fails, so just convert that into our usual "Out of memory allocating ..."
error

I didn't want to update all the callsites of chpl_glob to do this error
checking so I just made a glob_w wrapper that does the checking. That seemed
like a nice cleanup, so I made similar wrappers for the other glob routines to
cleanup a lot of safeCasts at callsites.

This also converts a halt() for bad zippered iteration to use the new
zipLengthHalt() wrapper.