update a comment
The code with this comment is due to 31a8a95ca47.
If we remove these checks as the comment suggests,
we will overlook those cases where our Symbol
gets modified, so it may be const-propagated incorrectly.

Conversely, the intents for 'this' argument have improved
since then, so maybe the intent checks will work.
Remember to check for the ref maybe-const intent, too.