Flag to indicate compiler-generated function's line number is reasonable
This change is motivated by default-initializers. If a field
initialization could not be resolved, like so:

  var x : int = 5.3;

The compiler would complain about the error at the site of the object's
creation (e.g. 'new') and not the field declaration.

This change adds a flag to indicate that the '_default_' wrapper
function's line number is acceptable.