Fix random indexing computation for bale histogram benchmark
I was capping the max index to M for an M*numTasks element array. M is the
number of entries in the table per-task, so I was accidentally capping the
index space to only the first few elements in the array. This corrects the
indexing and should help stabilize performance timings a little bit. It will
also improves performance when using BlockDist for the array (before this only
the atomics that lived on locale-0 were updated so it was effectively a
many-to-one test.)