Like PR #9970, this updates test output w.r.t. numa locale setup
We apparently don't test this for gasnet.numa, and the results came
back late enough on Friday that I hadn't noticed this case needed to
be updated as well.  Same reason as the others: the numa locale model
does a realloc and so gets parallel debugging printing for this test
requiring the output to be updated.