Preserve array shapes upon promotion (prototype)
This PR adds a _shape_ field to _iteratorRecord created during promotion
and uses it when creating an array out of it.
Note: as of this writing, it is not ready for merging.

Consider a promoted + on two arrays:

    var A = B + C;

During resolution, this is converted to:

    iter +(arg1,arg2) {
      forall (a,b) in zip(arg1,arg2) do
	yield a+b;
    }

    var temp = +(B,C);
    A = chpl_initCopy(temp);

where chpl_initCopy() is one the one in ChapelArray.chpl
that takes an _iteratorRecord and creates a 1-D array
out of the things that the corresponding iterator yields.

This PR makes resolution extend the _iteratorRecord type of the `temp`
with a _shape_ field and add the following statement:

    var temp = +(B,C);
    temp._shape_ = B.domain._instance;  <---------
    A = chpl_initCopy(temp);

chpl_initCopy() now checks whether the _iteratorRecord argument
has a meaningful shape. If so, it creates an array over that shape
instead of the 1-D array.

While ideally we would like to represent loops and loop expressions
using dedicated AST nodes, this PR achieves shape preservation
for promoted expressions with only small compiler modifications.

Test Changes
------------

* Un-futurized `test/arrays/tmacd/promotionLosesArrayShape`

This was the future requesting to preserve shapes.

* `test/distributions/robust/arithmetic/basics/test_promotion3`
  `test/studies/matrixlib/test_overloading`

Updated .good to shape-preserved output.

* Un-futurized `test/arrays/bradc/sliceViaSingleton`

"While there." Looks like it is producing the desired output now.

Next Steps
----------

* Preserve the shape of forall-expressions.

If we use the same approach, the challenge will be to generate an
appropriate `ir._shape_ = D` statement and insert it at the right spot.

Given that BenH needs a different representation of forall-expressions
to address initializer issues, it will is wise to do so first.

After that, the `ir._shape_ = D` statement will be easy to generate
during normalization.

* Once the above is done, update this PR and fix test failures.