Simplify tasking (and threading) implementation-specific .h file naming.
We have been naming the .h files for tasking layer implementations using
the form runtime/include/tasks/NAME/tasks-NAME.h, where NAME is the name
of the implementation ('qthreads', for example).  This required us to
parameterize references to these .h files such as in the general include
file runtime/include/chpl-tasks.h.  For this we used the preprocessor
variable CHPL_TASKS_MODEL_H which was defined in the runtime Makefiles.
To avoid implementation-defined aspects of the C standard, that variable
defined the entire double-quoted filename; that is, the double-quotes
were part of the variable string.  Unfortunately that exposed them to
quote-stripping if the variable happened to get passed through a shell
invocation.

Here, simplify things by adopting the convention that the implementation
specific .h file is named runtime/include/tasks/NAME/chpl-tasks-impl.h.
This allows us to use a fixed name in the general include directives,
while controlling which instance of the file is found during compilation
by means of the -I option we were already throwing.  (We use this same
technique for the comm layers.)

Do the same for CHPL_THREADS_MODEL_H also.