Build position independent third-party libraries when CHPL_LIBMODE=shared
This extends work done in #945 and permits our static third-party libraries to
be linked in by a shared object. .sos are PIC by default and require any static
libraries they use to be PIC as well. Previously we only built our runtime with
PIC support (so only quickstart was supported), and this extends that support
for all our third-party libraries as well.

Note that this is intended for use when Chapel is being used as a library and
is currently opt-in. We do not want to make it the default since there are
performance implications for PIC code.