Re-enable compile-time error messages for param int->enum casts
I'd turned off less of this code than I thought, so re-enabling this
error was not so hard.  While here, I made it a "continue" error
since we're generally trying to use those more.  I also added a test
that ensures that we don't rely  on the old-style automatic numbering
of enums.