Fix enum->string param casts
These broke because there was an 'if imm != NULL' check that no longer
held (since I don't convert enums to immediates as aggressively)
even though the imm wasn't used in that case.  I fixed this by pushing
the imm != NULL check to those specific cases that needed to refer to
'imm'.