Reduce dependence on CHPL_HOME in updateBuildVersion
It used to be that we required CHPL_HOME to be set when building
Chapel and, as a result, the updateBuildVersion script required it to
be set.  Over time, we've relaxed that requirement, but
updateBuildVersion still needed it to be set in order to avoid setting
the "SHA" part of the version number to -999.  This mod changes the
way it's defined to accept CHPL_HOME as a second argument.  Given that
the Makefile calling it has already determined where CHPL_HOME is and
can pass it in, this reduces the requirement that it be set by the user.