Use name grabbed from referenced type instead of modifying it myself (mostly)
We ran into problems with functions/gbt/retRecordMultiDef.chpl because it
exported a function which returned a record and the ref name did not include a
trailing "_chpl" while the record name itself did.  So, instead of adjusting
and hardcoding these relationships, take advantage of the fact I can determine
the referenced type's name from the referenced type itself and just modify that
to have a trailing "*".