Add a function to read multiple HDF5 files into a single 1-D array
Add a function `readNamedHDF5FilesInto1DArrayInt` to read all HDF5 files
listed in an argument into a single 1-D array.  This can be used to split
a large file into multiple smaller files to be read in parallel. It assumes
the data files are all the same size, with 2-D arrays of int(64)s.

Split `readAllHDF5Files()` into `readAllHDF5Files()` and
`readAllNamedHDF5Files()` to allow reading a list of files instead of always
reading all files in a directory.

Add a test that reads a 6000x5000 array (into 1-D) from 8 files.