Handle LoopExprs during scopeResolve
First, unify scope resolution for Expr* and conditionally pass off to
pre-existing functions.

Next, update 'getScope' to handle a parent LoopExpr. A special case is
the LoopExpr's iteratorExpr, whose scope is actually outside the
LoopExpr.

Finally, introduce code that creates a new scope when a LoopExpr is
encountered. A LoopExpr stores a flat list of DefExprs to be used as
indices, which we use to extend the newly-created scope. From there we
can resolve the other fields of the LoopExpr.