Lower LoopExprs during normalization
The bulk of the code added to LoopExpr.cpp is nearly-identical to the
original lowering in build.cpp.

Now that LoopExprs survive scope resolution, the indices for the
LoopExpr have actually been resolved to SymExprs. Alas, other parts of
the compiler (e.g. destructureIndices, loop-builders) are easier to work
with when using UnresolvedSymExprs as they will create DefExprs in the
right spots. In order to leverage this, replace the indices with
UnresolvedSymExprs.

LoopExprs need to be handled differently depending on whether they are
inside an ArgSymbol. For LoopExprs at module scope or within a function,
they can be lowered in the same manner as they used to: create several
nested functions at the scope of the loop-expr. This doesn't work for
LoopExprs inside ArgSymbols because that pattern is generally not
supported by resolution. There are issues with visible-functions,
scoped-types disappearing, and other bizarre unhandled behavior.

Instead, we will generate a single function at module scope and pass any
required data to that function. The various iterator functions will be
nested within this top-level function in order to share visibility for
those outer variables.

Finally, each function will go through the same post-processing:
- add function to the symbol table
- lower short-circuit operators
- scope resolution
- normalization

The core normalize pass is updated to handle a LoopExpr as a variable's
initialization expression and to lower LoopExprs as part of arbitrary
AST normalization. Lowering of IfExprs and short-circuit operators is
tweaked to return early if the expression is inside a LoopExpr, in favor
of returning to those expressions once the LoopExpr is lowered.