Adjust analyzeMemLeaksLog to handle the new memStats format
We adjusted `--memStats` output in #10061 and that broke analyzeMemLeaksLog,
which looks for sentinels and specific strings. In order to get `--memStats`
working with multi-locale we removed the header/footer that analyzeMemLeaksLog
looked for, and we changed some of the wording. This ajdusts analyzeMemLeaksLog
to stop looking for the header/footer and to expect the new wording.